+++ src/components/Header/index.css
@@ -1,0 +1,28 @@
+.scroll-overpass {
+  @apply bg-secondary-100/60
+    dark:bg-secondary-900/60
+    backdrop-blur-md shadow-md;
+}
+
+.scroll-overpass-hide {
+  @apply bg-transparent dark:bg-transparent backdrop-blur-none shadow-none;
+}
+
+.scroll-overpass-hide > * {
+  @apply opacity-0;
+}
+
+header {
+  @apply fixed z-50 top-0 p-4
+    flex flex-col
+    bg-transparent
+    dark:bg-transparent
+    sm:static
+    sm:flex-row
+    sm:justify-between
+    sm:h-app-header;
+}
+
+header button.menu {
+  @apply m-0 p-0  text-4xl bg-transparent shadow-none;
+}
