+++ src/components/Menu/index.css
@@ -1,0 +1,56 @@
+nav.menu {
+  @apply fixed z-[100] top-0 w-full h-screen p-8
+    flex flex-col items-center justify-center
+    bg-secondary-100
+    dark:bg-secondary-900
+    pointer-events-none
+    opacity-0
+    sm:static
+    sm:w-fit
+    sm:h-auto
+    sm:opacity-100
+    sm:p-0
+    sm:flex-row
+    sm:bg-transparent
+    sm:dark:bg-transparent
+    sm:pointer-events-auto;
+}
+
+nav.menu-show {
+  @apply opacity-100 animate-slide-in-x pointer-events-auto;
+}
+
+nav.menu-hide {
+  @apply opacity-0 animate-slide-out-x;
+}
+
+nav.menu ul {
+  @apply flex flex-col sm:flex-row justify-center;
+}
+
+nav.menu li {
+  @apply flex items-center justify-center mt-8 sm:mt-0 sm:ml-2 lg:ml-8 text-xl sm:text-base cursor-pointer;
+}
+
+nav.menu li.brand {
+  @apply mb-8 text-4xl font-bold;
+}
+
+nav.menu li a, nav.menu li div {
+  @apply no-underline px-4 py-2 rounded-full;
+}
+
+nav.menu li a:hover, nav.menu li div:hover {
+  @apply bg-selected-200 text-selected-900
+    dark:bg-selected-600 dark:text-white;
+}
+
+nav.menu li a:active, nav.menu li div:active {
+  @apply transform scale-95;
+}
+
+nav.menu li .selected {
+  @apply bg-selected-100 text-selected-900
+    dark:bg-selected-300 dark:text-selected-900 
+    shadow-md;
+}
