+++ src/components/Vaults/index.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import {FilterProvider} from './useFilter';
-import Header from './Header';
 import List from './List';
+import Filter from './Filter';
 import './index.css';
 
 export default function Vaults() {
-	return <div className={'vaults'}>
-		<FilterProvider>
-			<Header></Header>
-			<List />
-		</FilterProvider>
-	</div>;
+	return <FilterProvider>
+		<Filter></Filter>
+		<List />
+	</FilterProvider>;
 }
